First major update
Added a manager that has tons of public method for other mod to use.
Added ability to create new music entry.
Added ability to create new SFX entry.
Added ability to load file relative to the mod path (for example: BattleMusic_Titans = Hearts Stay Unchanged.mp3) is a valid path if there's a file with this name next to the mod dll.
Added ability to access the audio metadata (tested on id3v1, and id3v2) using ATL 3.25, will not work if the metadata is stripped.
Added a wrapper SoundData to easily access the entry/definition/audioclip/metadata of your sound.
Scrollbar asset bundle will now not be loaded if the in-game menu is disabled.
Playing an SFX from the mod menu will now temporarily disable looping for that SFX.